Implement variable staging and creation for L1 agent
The L1 IPOMDP uses the parsed variable names as the state space and
observation space. In addition, a separate variable for agent j's model,
M_j is added to the state vars to create interactive state space. J's
observations are added to i's set of obsVars to track them depending on
physical state S.

Once the state and obsVars are figured out, DDs can be created for M_j
transitions and joint action transitions