Change Opponent model to traverse policy trees iteratively
The Opponent Model variable M_j has an arity of number of nodes in the
opponent model. Building a transition DD out of this takes a lot of
memory due to the huge branching factor of some policy trees. Instead,
The Opponent Model up to a certain number of maximum horizon is built
offline after solving the lower frames. The Opponent Model is then
traversed partially up to a fixed horizon and the transition function of
the visited nodes is built on the fly. This creates a potentially
dynamic state space which does not allow the formation of an alpha
vector policy. However, it may help with overcoming problems of state
space explosion and policy trees can be used instead as a policy
representation.

TODO: Test new changes to Opponent Model API