Implement Mj look ahead for dynamic IS creation
The IPOMDP class implements dynamic Mj creation along with creation of
Ti, Oi, and Oj DDs. The previous data structure which used a HashMap of
HashMaps to store these function has been changed to use a HashMap of DD
arrays for more efficient access of complete arrays.

THe indices of primed and unprimed variables are recorded to ease
OP.addMultVarElim in the future while implementing belief updates.