Condition Ri on Mj for every IS change
Build reward function using the setDynamics mechanism of POMDPs.
Additionally, include Mj as a condition for Ri.

IPOMDPParser overrides the mechanism to construct R since the orignal
ParseSPUDD method stores the costs as a vector corresponding to the
action index. This makes it difficult to access it using joint action
names for POMDPs. Instead, map to the joint action names in a hashmap
and do everything else same as the ParseSPUDD implementation