Refactor and rename to follow naming conventions
The original POMDP solver code is moved to a directory named legacy
code. This sub package still contains Hoey's core implementation of DDs
and their operations and should not be messed with. Other code is moved
outside and away from the legacy code to avoid unintended changes to the
core implementation