Remove unused imports in DDTree classes and rename the package
The DDTree implementation originally written to create SPUDD files is
now used in the POMDP and IPOMDP classes to hold non global DDs which
don't use variables in the current context. These are moved to a
separate package. The unused imports in those classes are removed and
proper headers are added