Fix logging, fix IPOMDP IS init and add CLI args to BUViewer
Some of the classes in the solver logged full complete DD structures.
This made the logs unreadable. Instead, only the variables in the factor
represented by the DD are now printed. Also, the variable committing in
IPOMDPs was redundant (once while initing online functions and again
while initing offline functions). This is fixed.