Fix initial policy in SSGA belief expansion
The SSGA expansion object needs a policy to search into the belief
space. Previously, a full belief expansion was done first for just 1
time step to gather some beliefs for creating a policy. But this
resulted in a lot of beliefs that were never reached because of the
roots from the full belief expansion. Now, the reward function is taken
as a default policy to begin searches into the belief space