Fix belief tree root tracking in multiple lower level frames
The dynamic belief tree keeps a track of the next level nodes after
expanding from roots during each iteration. This is so that stepping can
be verified using the non zero nodes obtained from the belief update.
For multiple frames, the verification was including nodes of other
frames which used to cause out of sync errors.