Fix ordering of obsCombinations
The obsCombinations for both POMDPs and IPOMDPs have to be generated
according to the sequence from the statedecode API in the legacy code.
This is because all the other functions such as convert2array use that
sequence by default. So having a different sequence will lead to
incorrect flattening of DDs. This is fixed by using the statedecode API
to generate both obsCombinations and the indices of their values.