Fix conditional plan printing
When solving lower frames and switching between their contexts while
expanding Mj look ahead trees, the globals for IPOMDP are over written.
This is not undone as soon as the execution switches back to the IPOMDP.
This is fixed. Also, joint action transitions now assign 0.99
probability to the most probable action by agent i.