Use joint distributions for POMDPs
The POMDP solver was largely based on exploiting factors of the joint
belief. However, in our case, one of the domains had an instance where
the state variables were not independent of each other. So I rewrote the
DPbackup and NextBelState computation in terms of joint distribution. I
do not know if this is a good or a bad idea or if it even solves the
problem. But It does pass all test cases. So I am going with it